feat: introduce API-only architecture and enhance authentication flow
- Added a new `auth.sh` script for simplified Discord login process, providing users with a direct URL for authentication.
- Updated `CLAUDE.md` to reflect the transition to an API-only version of Vinoflare, detailing essential commands and API endpoints.
- Created `MIGRATION_TO_API_ONLY.md` to guide developers through converting a full-stack application to an API-only service.
- Removed frontend components and related configurations, streamlining the project structure for backend-only functionality.
- Updated `README.md` to emphasize the API-focused nature of the project and included instructions for API integration.
- Refactored `package.json` and `vite.config.ts` to remove client-specific configurations and dependencies.

These changes establish a clear API-only architecture, enhancing the project's focus on backend services and improving developer experience.
