+++ README.md
@@ -1,16 +1,17 @@
-# Vinoflare v2 - Modern Full-Stack Template for Cloudflare Workers
+# Vinoflare v2 API - Modern API Template for Cloudflare Workers
 
-一个基于 Cloudflare Workers 的现代全栈应用模板，集成了 React、Hono、Drizzle ORM 和完整的类型安全体系。
+一个基于 Cloudflare Workers 的现代 API 服务模板，集成了 Hono、Drizzle ORM、Better Auth 和完整的类型安全体系。
 
 ## 🚀 特性
 
 - **边缘优先**: 基于 Cloudflare Workers，享受全球边缘网络的性能优势
-- **类型安全**: 从数据库到 API 到前端的端到端类型安全
+- **类型安全**: 从数据库到 API 的端到端类型安全
 - **模块化架构**: 清晰的模块化设计，易于扩展和维护
-- **自动化代码生成**: 一键生成 CRUD 模块、API 客户端和类型定义
-- **现代技术栈**: React 19 + TypeScript + Vite + TanStack Router
+- **自动化代码生成**: 一键生成 CRUD 模块和类型定义
+- **现代技术栈**: Hono + TypeScript + Drizzle ORM
 - **身份认证**: 集成 Better Auth，支持 Discord OAuth
 - **API 文档**: 自动生成 OpenAPI 文档和交互式 UI
+- **API 专注**: 纯 API 服务，无前端代码，适合作为微服务或 BFF
 
 ## 📋 前置要求
 
@@ -58,18 +59,17 @@ bun run db:push:local
 bun run dev
 ```
 
-访问 http://localhost:5173 查看应用
+API 服务将在 http://localhost:5173 启动
+
+- API 文档：http://localhost:5173/api/docs
+- 健康检查：http://localhost:5173/api/health
+- OpenAPI 规范：http://localhost:5173/api/openapi.json
 
 ## 🏗️ 项目结构
 
 ```
 src/
-├── client/               # React 前端应用
-│   ├── components/       # UI 组件
-│   ├── hooks/           # 自定义 Hooks 和 API 集成
-│   ├── lib/             # 客户端工具库
-│   └── routes/          # 页面路由（TanStack Router）
-├── server/              # Hono 后端应用
+├── server/              # API 服务端代码
 │   ├── config/          # 配置文件
 │   ├── core/            # 核心功能（模块加载、错误处理）
 │   ├── db/              # 数据库表结构和迁移
@@ -79,7 +79,8 @@ src/
 │   ├── routes/          # 路由配置
 │   ├── schemas/         # Zod 验证模式
 │   └── types/           # TypeScript 类型定义
-└── generated/           # 自动生成的代码
+├── generated/           # 自动生成的代码
+└── index.ts            # 应用入口
 ```
 
 ## 🔧 常用命令
@@ -95,19 +96,14 @@ bun run lint:fix         # 自动修复代码问题
 bun run test             # 运行测试（使用 Vitest）
 ```
 
-> ⚠️ **注意**: 请使用 `bun run test` 而不是 `bun test`。本项目使用 Vitest 配合 Cloudflare Workers 进行测试，需要特殊配置。
-
 ### 代码生成
 
 ```bash
 # 生成新的 CRUD 模块（推荐）
-bun run scaffold:module <模块名>
+bun run gen:module <模块名>
 
-# 生成 API 客户端和类型
+# 生成 OpenAPI 规范
 bun run gen:api
-
-# 生成路由类型
-bun run gen:routes
 
 # 生成 Cloudflare 绑定类型
 bun run gen:types
@@ -127,7 +123,7 @@ bun run db:studio        # 打开数据库管理界面
 使用脚手架快速创建新模块：
 
 ```bash
-bun run scaffold:module products
+bun run gen:module products
 ```
 
 这会自动生成：
@@ -141,14 +137,14 @@ bun run scaffold:module products
 生成后需要：
 1. 运行 `bun run db:generate` 生成迁移
 2. 运行 `bun run db:push:local` 应用迁移
-3. 运行 `bun run gen:api` 更新客户端类型
+3. 运行 `bun run gen:api` 更新 OpenAPI 规范
 
 ## 🔐 身份认证
 
 项目集成了 Better Auth，默认配置了 Discord OAuth：
 
 1. 在 [Discord Developer Portal](https://discord.com/developers/applications) 创建应用
-2. 添加重定向 URL：`http://localhost:5173/api/auth/callback/discord`
+2. 添加重定向 URL：`http://localhost:5173/api/auth/callback/discord`（生产环境使用实际域名）
 3. 将 Client ID 和 Secret 添加到 `.dev.vars`
 
 支持的认证功能：
@@ -156,6 +152,29 @@ bun run scaffold:module products
 - JWT 令牌
 - 会话管理（7天有效期）
 - 路由保护（默认所有 API 路由需要认证）
+
+### 简单登录
+
+```bash
+# 获取 Discord 登录 URL
+./auth.sh
+
+# 在浏览器中打开显示的 URL 完成登录
+# 登录后浏览器会自动保存 session cookie
+
+# 退出登录：在浏览器中清除 localhost:5173 的 cookies 即可
+```
+
+详细使用说明请参考 [API_AUTH_GUIDE.md](./API_AUTH_GUIDE.md)
+
+### 公开路由
+
+以下路由无需认证即可访问（在 `src/server/config/routes.ts` 中配置）：
+- `/api/hello` - 测试端点
+- `/api/auth/*` - 认证相关端点
+- `/api/openapi.json` - API 规范
+- `/api/docs` - API 文档
+- `/api/health` - 健康检查
 
 ## 📝 API 文档
 
@@ -213,7 +232,7 @@ bun run deploy
 ### 类型安全
 - 数据库模式是类型的唯一来源
 - 使用 `drizzle-zod` 从数据库模式生成验证
-- 通过 Orval 生成类型安全的 API 客户端
+- 导出类型供客户端使用
 
 ### 错误处理
 - 使用 `HTTPException` 处理标准 HTTP 错误
@@ -222,8 +241,36 @@ bun run deploy
 
 ### 性能优化
 - 利用 Cloudflare Workers 的边缘计算能力
-- 使用 Vite 的代码分割优化包大小
 - 实施适当的缓存策略
+- 使用连接池优化数据库访问
+
+## 🔗 与前端集成
+
+这是一个纯 API 项目，可以与任何前端框架集成：
+
+### 使用 OpenAPI 客户端生成器
+```bash
+# 使用 openapi-typescript-codegen
+npx openapi-typescript-codegen --input http://localhost:5173/api/openapi.json --output ./src/api
+
+# 或使用 orval
+npx orval --input http://localhost:5173/api/openapi.json --output ./src/api
+```
+
+### 手动集成示例
+```typescript
+// 获取认证会话
+const response = await fetch('http://localhost:5173/api/auth/get-session', {
+  credentials: 'include'
+});
+
+// 调用受保护的 API
+const todos = await fetch('http://localhost:5173/api/todos', {
+  headers: {
+    'Authorization': `Bearer ${session.token}`
+  }
+});
+```
 
 ## 🤝 贡献
