+++ package.json
@@ -4,12 +4,11 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "rm -rf dist && vite build --mode client && vite build",
+    "build": "rm -rf dist && vite build",
     "deploy": "wrangler deploy",
     "typecheck": "tsc --noEmit",
-    "gen:routes": "tsr generate",
     "gen:types": "wrangler types --env-interface CloudflareBindings",
-    "gen:api": "bun scripts/generate-openapi.ts && bun orval --config orval.config.ts",
+    "gen:api": "bun scripts/generate-openapi.ts",
     "gen:module": "bun scripts/generate-module/index.ts",
     "db:generate": "drizzle-kit generate",
     "db:push:local": "wrangler d1 migrations apply DB --local",
@@ -30,17 +29,10 @@
     "@cloudflare/vitest-pool-workers": "^0.8.46",
     "@cloudflare/workers-types": "^4.20250620.0",
     "@hono/node-server": "^1.14.4",
-    "@tailwindcss/postcss": "^4.1.10",
-    "@tanstack/router-cli": "^1.121.34",
-    "@tanstack/router-plugin": "^1.121.34",
     "@types/node": "^24.0.4",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
     "better-sqlite3": "^12.1.1",
     "dotenv": "^16.5.0",
     "drizzle-kit": "^0.31.2",
-    "orval": "^7.10.0",
-    "postcss": "^8.5.6",
     "typescript": "^5.8.3",
     "vite": "6.3.5",
     "vitest": "^3.2.4",
@@ -48,18 +40,11 @@
   },
   "dependencies": {
     "@scalar/hono-api-reference": "^0.9.7",
-    "@tailwindcss/vite": "^4.1.10",
-    "@tanstack/react-query": "^5.81.2",
-    "@tanstack/react-router": "^1.121.34",
-    "axios": "^1.10.0",
     "better-auth": "^1.2.10",
-    "clsx": "^2.1.1",
     "drizzle-orm": "^0.44.2",
     "drizzle-zod": "^0.8.2",
     "hono": "^4.7.3",
     "http-status-codes": "^2.3.0",
-    "sonner": "^2.0.5",
-    "tailwind-merge": "^3.3.1",
     "zod": "^3.25.2"
   }
 }
