--- src/client/components/auth/login-form.tsx
@@ -1,62 +1,0 @@
-import { signIn } from "@client/lib/auth";
-import { useLocation } from "@tanstack/react-router";
-import { cn, colors, focus, interactive, text } from "@/client/lib/design";
-
-export function LoginForm() {
-	const location = useLocation();
-	const searchParams = new URLSearchParams(location.search);
-	const error = searchParams.get("error");
-
-	const handleSignIn = async () => {
-		try {
-			await signIn.social({ provider: "discord" });
-		} catch (error) {
-			console.error("Login error:", error);
-		}
-	};
-
-	return (
-		<div className="mx-auto w-full max-w-sm">
-			{error && (
-				<div className="mb-8 bg-red-50 p-4">
-					<p className={cn(text.small, "text-red-600")}>
-						{error === "OAuthCallbackError"
-							? "Authentication failed. Please try again."
-							: "An error occurred during sign in."}
-					</p>
-				</div>
-			)}
-
-			<div className="space-y-6">
-				<div>
-					<h1 className={cn(text.h1, "mb-2")}>Sign In</h1>
-					<p className={cn(text.base, colors.text.secondary)}>
-						Continue with your Discord account
-					</p>
-				</div>
-
-				<button
-					type="button"
-					onClick={handleSignIn}
-					className={cn(
-						"flex w-full items-center justify-center gap-3 px-4 py-3",
-						colors.bg.inverse,
-						colors.text.inverse,
-						interactive.base,
-						focus,
-					)}
-				>
-					<svg
-						className="h-5 w-5"
-						fill="currentColor"
-						viewBox="0 0 24 24"
-						aria-hidden="true"
-					>
-						<path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515a.074.074 0 0 0-.079.037c-.21.375-.444.864-.608 1.25a18.27 18.27 0 0 0-5.487 0a12.64 12.64 0 0 0-.617-1.25a.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057a19.9 19.9 0 0 0 5.993 3.03a.078.078 0 0 0 .084-.028a14.09 14.09 0 0 0 1.226-1.994a.076.076 0 0 0-.041-.106a13.107 13.107 0 0 1-1.872-.892a.077.077 0 0 1-.008-.128a10.2 10.2 0 0 0 .372-.292a.074.074 0 0 1 .077-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .078.01c.12.098.246.198.373.292a.077.077 0 0 1-.006.127a12.299 12.299 0 0 1-1.873.892a.077.077 0 0 0-.041.107c.36.698.772 1.362 1.225 1.993a.076.076 0 0 0 .084.028a19.839 19.839 0 0 0 6.002-3.03a.077.077 0 0 0 .032-.054c.5-5.177-.838-9.674-3.549-13.66a.061.061 0 0 0-.031-.03zM8.02 15.33c-1.183 0-2.157-1.085-2.157-2.419c0-1.333.956-2.419 2.157-2.419c1.21 0 2.176 1.096 2.157 2.42c0 1.333-.956 2.418-2.157 2.418zm7.975 0c-1.183 0-2.157-1.085-2.157-2.419c0-1.333.955-2.419 2.157-2.419c1.21 0 2.176 1.096 2.157 2.42c0 1.333-.946 2.418-2.157 2.418z" />
-					</svg>
-					<span className={text.base}>Continue with Discord</span>
-				</button>
-			</div>
-		</div>
-	);
-}
