--- src/client/routes/profile.tsx
@@ -1,10 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { UserProfile } from "../components/auth/user-profile";
-
-export const Route = createFileRoute("/profile")({
-	component: ProfilePage,
-});
-
-function ProfilePage() {
-	return <UserProfile />;
-}
