+++ tsconfig.json
@@ -2,8 +2,7 @@
 	"compilerOptions": {
 		"target": "ES2022",
 		"module": "ESNext",
-		"lib": ["ES2022", "DOM", "DOM.Iterable"],
-		"jsx": "react-jsx",
+		"lib": ["ES2022"],
 		"moduleResolution": "bundler",
 		"allowImportingTsExtensions": true,
 		"resolveJsonModule": true,
@@ -14,10 +13,9 @@
 		"noUnusedParameters": true,
 		"noFallthroughCasesInSwitch": true,
 		"skipLibCheck": true,
-		"types": ["@cloudflare/workers-types", "vite/client", "node"],
+		"types": ["@cloudflare/workers-types", "node", "vite/client"],
 		"paths": {
 			"@/*": ["./src/*"],
-			"@client/*": ["./src/client/*"],
 			"@server/*": ["./src/server/*"]
 		}
 	},
