--- tsr.config.json
@@ -1,4 +1,0 @@
-{
-	"routesDirectory": "./src/client/routes",
-	"generatedRouteTree": "./src/generated/routeTree.gen.ts"
-}
